Documents/Booking_Flight/flight_booking/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-rest-registration==0.7.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
